requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
alembic
python-multipart
aiofiles
python-jose
